+++ src/pages/CardRegister/CardRegister.style.ts
@@ -1,14 +1,5 @@
 import styled from 'styled-components';
 
-export const Container = styled.div`
-  width: 382px;
-  height: 700px;
-  background-color: #fff;
-
-  border-radius: 8px;
-  border: 3px solid #000;
-`;
-
 export const SubContainer = styled.div`
   width: 100%;
   height: 642px;
