+++ src/pages/RegisterConfirm/RegisterConfirm.tsx
@@ -1,0 +1,3 @@
+export default function RegisterConfirm() {
+  return <div>RegisterConfirm</div>;
+}
